arch: riscv: smp: allow custom IPI implementation
The currently IPI implementation assumes that CLINT exists in the
system, however, that might not be the case as IPI can be implemented
with certain PLIC as well, such as Andes NCEPLIC100 which supports
soft-triggering.

Refactor the CLINT-based IPI implementations into `ipi_clint.h`, and
create Kconfig that selects the CLINT implementation when
`sifive-clint0` exists and enabled, otherwise default to
`RISCV_SMP_IPI_CUSTOM` which allows a OOT implementation.
